python-service/app/schemas/sample_item.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict, Field


class SampleItemCreate(BaseModel):
    name: str = Field(min_length=1, max_length=120)
    description: str | None = Field(default=None, max_length=500)


class SampleItemRead(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    id: int
    name: str
    description: str | None
    created_at: datetime
